코드 분리 및 클래스로 이동
