+++ src/repository/room.ts
@@ -1,4 +1,6 @@
 import {Router} from "mediasoup/node/lib/Router";
+import {mediaCodecs} from "../constant/config.js";
+import {Worker} from "mediasoup/node/lib/Worker.js";
 
 export interface Room {
     router: Router;
@@ -7,6 +9,31 @@ export interface Room {
 
 class RoomRepository {
     #rooms: Map<string, Room> = new Map<string, Room>();
+
+    createRoom = async (roomName: string, socketId: string, worker: Worker) => {
+        // worker.createRouter(options)
+        // options = { mediaCodecs, appData }
+        // mediaCodecs -> defined above
+        // appData -> custom application data - we are not supplying any
+        // none of the two are required
+        let router: Router;
+        let socketIds: string[] = [];
+        const room = this.getRoomByName(roomName);
+
+        if (room !== undefined) {
+            router = room.router;
+            socketIds = room.socketIds || [];
+        } else {
+            router = await worker.createRouter({mediaCodecs});
+        }
+        console.log(`Router ID: ${router.id}`, socketIds.length);
+        this.setRoom(roomName, {
+            router: router,
+            socketIds: [...socketIds, socketId],
+        });
+
+        return router;
+    };
 
     getRoomByName = (roomName: string): Room | undefined => {
         return this.#rooms.get(roomName);
